feat: separate DetailView into heading and body
